// platform/frontend/src/client/types.gen.ts
export type AgentSessionPublic = {
    /**
     * Id
     */
    id: string;
    /**
     * Agent Key
     */
    agent_key: string;
    /**
     * Scope Type
     */
    scope_type: string;
    /**
     * Scope Id
     */
    scope_id: string;
    /**
     * Title
     */
    title: string;
    /**
     * Status
     */
    status: string;
    /**
     * Created At
     */
    created_at?: string | null;
    /**
     * Updated At
     */
    updated_at?: string | null;
};